pyproject.toml
[project]
name = "listonic-ha"
version = "0.1.0"
description = "Home Assistant custom component for Listonic shopping lists"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "homeassistant>=2024.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.0",
    "python-dotenv>=1.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_components/listonic"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
